Add rest of tools packages to compile script
The compile script, in it's current form, is done. All it does is extract
and make packages, saving non-error output to a log file and printing
the error output to the screen (I liked being able to see when something
messed up), but it needs to be worked on. For instance, I haven't ran
the whole thing all at once, and I also would like to do things like:
Make a log folder with the date, for nice organization
Add in the make checks
Save the error output to the log file
Save make check output to a specific file
Separate config and make output
More nicely split up the tasks
Make sure nothing else gets printed while doing the tasks
But for now, there's nothing wrong with it. It works, it's just not the
best thing I could have made.