add(script): compilation helper
This script should help with compilation a lot. Ideally, I make a
"package" file for each progam I want to install, with a function for
each important step. Then I can use this wrapper function to take care
of the error printing and quitting, and logging all the output and
everything. I have yet to change anything to use it, but it's here.